pyproject.toml
[tool.poetry]
name = "youtube-summarizer"
version = "0.1.0"
description = "A YouTube video summarizer using OpenRouter with DeepSeek V3"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
youtube-transcript-api = "^0.6.1"
groq = "^0.4.2"
openai = "^1.12.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
summarize = "youtube_summarizer.cli:main"
